Added gameplay
Some things has been added to granted the game with more funny stuffs.
- A bug will appear in a random place of the game board, giving you
points if catch it.
- A snake will appear like the bug, but only in the place between the
road and the water. If it touchs you, you're dead
- A loop game where the goal is to reach the maximum number of points.
You achieve this task, you count with 3 lifes. When you reachs the top
of the gameboard, an animation is displays and 100 points are given to
you.